fix(gsc): share PEM loader with google-auth and mock in unit tests
Centralize key normalization in src/lib/google-sa-key.ts so GSC/Calendar
suites mock loadGooglePrivateKey instead of the old jose importPKCS8 path.
